Factor our message handling. Would be better to share with the manager side.
